% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDNAoccupancy_stan.R
\name{eDNAoccupancy_stan}
\alias{eDNAoccupancy_stan}
\title{Three-level occupancy model}
\usage{
eDNAoccupancy_stan(n_sites, n_Psi_coef, X_psi, n_samples, n_p_coef,
  n_Theta_coef, V_p, W_theta, y, a_obs, k, start_index, end_index, z_obs,
  n_samples_per_site, ...)
}
\arguments{
\item{n_sites}{Number of sites}

\item{n_Psi_coef}{Number of Psi coefficients}

\item{X_psi}{matrix of psi predictors}

\item{n_samples}{total number of samples}

\item{n_p_coef}{number of p coefficients}

\item{n_Theta_coef}{number of theta coefficients}

\item{V_p}{matrix of p predictors}

\item{W_theta}{matrix of theta predictors}

\item{y}{vector of number of detections per sample}

\item{a_obs}{vetor if detections occured within sample MTI}

\item{k}{number of molecular replicates per sample}

\item{start_index}{starting index for each site used MTI}

\item{end_index}{ending index for each site used MTI}

\item{z_obs}{vector for each site if any detections occured}

\item{n_samples_per_site}{number of samples per each site}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Parameter with MTI will likely be moved to internal functions
in future version of this function
}
\details{
Also, site needs to be changed to sampling event
}
