% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_stan.R
\name{logistic_stan}
\alias{logistic_stan}
\title{Simple bayesian logistic regression with Stan}
\usage{
logistic_stan(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Simple bayesian logistic regression with Stan
}
\examples{
2 + 2
}
