% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USFWS_UMR_eDNA.R
\name{USFWS_UMR_eDNA_stan}
\alias{USFWS_UMR_eDNA_stan}
\title{Occupancy model used by USFWS for eDNA on UMR}
\usage{
USFWS_UMR_eDNA_stan(n_obs, n_sites, n_psi_coef, X_psi, site_detections,
  n_samples, n_theta_coef, V_p, AC1, AC3, start_index, end_index,
  any_detected, k, n_row_sep, n_col_sep, site_event_predict,
  n_predict_samples, delta_sd, alpha_sd, beta_sd, delta_mean, alpha_mean,
  beta_mean, ...)
}
\arguments{
\item{n_obs}{number of total observations}

\item{n_sites}{number of sampling sites (e.g., pools)}

\item{n_psi_coef}{number of site-level coefficients}

\item{X_psi}{matrix for site-level predictors}

\item{site_detections}{binar vector of eDNA detetion at each site}

\item{n_samples}{vector of no. samples per site}

\item{n_theta_coef}{number of sample-level coeffients}

\item{V_p}{matrix for sample-level predictors}

\item{AC1}{number of positive detections per sample using ACTM 1 assasy}

\item{AC3}{number of positive detections per sample using ACTM 3 assasy}

\item{start_index}{sample-level information}

\item{end_index}{sample-level information}

\item{any_detected}{were there any_detected in sample? 0 or 1}

\item{k}{number of molecular replicates per sample}

\item{n_row_sep}{number of rows in predictor matrix}

\item{n_col_sep}{number of columns in predicotr matrix}

\item{site_event_predict}{matrix of site-event predictions}

\item{n_predict_samples}{number of prediciton samples}

\item{delta_sd}{prior for delta SD}

\item{alpha_sd}{prior for alpha SD}

\item{beta_sd}{prior for beta SD}

\item{delta_mean}{prior for delta mean}

\item{alpha_mean}{prior for alpha mean}

\item{beta_mean}{prior for beta mean}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}

\item{W_theta}{matrix for sample-level predictors}

\item{n_p_coef}{number of detection-level coeffients}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
This model is was used by the USFWS on the UMR as part
of their bighead and silver carp monitoring.
}
\details{
This will include a reference to Mize et al. once the
help file is fleshed out.

With this model, I will need to move more details into hte r
}
\examples{
2 + 2
}
