% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_stan.R
\name{logistic_stan}
\alias{logistic_stan}
\title{Simple bayesian logistic regression with Stan}
\usage{
logistic_stan(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Simple bayesian logistic regression with Stan
}
\examples{
set.seed(12345)
n_reps = 10000
y <- rbinom(n = n_reps * 2, 1, c(0.25, 0.75))
x <- rep(c(0, 1), times = n_reps)

l_out <- logistic_stan(y, x)

## show simulated data is almost the same as the recovered probs
print(rstan::summary(l_out)$summary[ ,"mean"][1:2])
print(aggregate(y, by = list(x), mean)$x)


}
