% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_target_stan.R
\name{logistic_target_stan}
\alias{logistic_target_stan}
\title{Bayesian logistic regression using traget probabiltiy with Stan}
\usage{
logistic_target_stan(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
This differes from \code{logistic_stan()} becaues of how the
Stan code is written. Both this function and that function
are included as teaching tools to build up towards occupancy models.
}
\examples{
2+2 
}
